+++ dependencies.py
@@ -40,7 +40,7 @@ Dependency = namedtuple("Dependency", ["name", "package", "constraint"])
 DEPENDENCIES = [
     Dependency("SQLAlchemy", "sqlalchemy", ">=1.1.11, <1.2"),
     Dependency("GeoAlchemy2", "geoalchemy2", ">=0.6.2, <0.7"),
-    Dependency("lizard-connector", "lizard_connector", "==0.6"),
+    Dependency("lizard-connector", "lizard_connector", "==0.7.3"),
     Dependency("pyqtgraph", "pyqtgraph", ">=0.10.0"),
     Dependency("threedigrid", "threedigrid", "==1.0.24"),
     Dependency("cached-property", "cached_property", ""),